+++ react/components/FilterSidebar.js
@@ -1,6 +1,6 @@
 import classNames from 'classnames'
 import produce from 'immer'
-import React, { useState, useEffect, useMemo, Fragment } from 'react'
+import React, { useState, useEffect, useMemo, Fragment, useRef } from 'react'
 import { FormattedMessage } from 'react-intl'
 import { ExtensionPoint } from 'vtex.render-runtime'
 import { Button } from 'vtex.styleguide'
@@ -12,10 +12,15 @@ import FilterNavigatorContext, {
 } from './FilterNavigatorContext'
 import AccordionFilterContainer from './AccordionFilterContainer'
 import Sidebar from './SideBar'
+import { MAP_CATEGORY_CHAR } from '../constants'
 import { buildNewQueryMap } from '../hooks/useFacetNavigation'
 
 import styles from '../searchResult.css'
 import { getFullText } from '../utils/compatibilityLayer'
+import {
+  isCategoryDepartmentOrCollection,
+  filterCategoryDepartmentAndCollection,
+} from '../utils/queryAndMapUtils'
 
 const CSS_HANDLES = [
   'filterPopupButton',
@@ -37,6 +42,7 @@ const FilterSidebar = ({
   const filterContext = useFilterNavigator()
   const [open, setOpen] = useState(false)
   const handles = useCssHandles(CSS_HANDLES)
+  const shouldClear = useRef(false)
 
   const [filterOperations, setFilterOperations] = useState([])
   const [categoryTreeOperations, setCategoryTreeOperations] = useState([])
@@ -75,12 +81,18 @@ const FilterSidebar = ({
   }
 
   const handleClearFilters = () => {
-    /* This navigateToFacet is a hack to solve the problem of this
-     function not updating the checkbox status, giving the impression
-     that the 'clean' button was not working */
-    navigateToFacet([], preventRouteChange)
-    setFilterOperations([])
-    setOpen(false)
+    shouldClear.current = true
+    // Gets the previously selected facets that should be cleared
+    const selectedFacets = selectedFilters.filter(
+      facet => !isCategoryDepartmentOrCollection(facet.key) && facet.selected
+    )
+
+    // Should not clear categories, departments and clusterIds
+    const selectedRest = filterOperations.filter(facet =>
+      isCategoryDepartmentOrCollection(facet.key)
+    )
+
+    setFilterOperations([...selectedFacets, ...selectedRest])
   }
 
   const handleUpdateCategories = maybeCategories => {
@@ -88,7 +100,11 @@ const FilterSidebar = ({
       ? maybeCategories
       : [maybeCategories]
 
-    const categoriesSelected = filterOperations.filter(op => op.map === 'c')
+    /* There is no need to compare with CATEGORY and DEPARTMENT since
+     they are seen as a normal facet in the new VTEX search */
+    const categoriesSelected = filterOperations.filter(
+      op => op.map === MAP_CATEGORY_CHAR
+    )
     const newCategories = [...categoriesSelected, ...categories]
 
     // Just save the newest operation here to be recorded at the category tree hook and update the tree
@@ -97,7 +113,7 @@ const FilterSidebar = ({
     // Save all filters along with the new categories, appended to the old ones
     setFilterOperations(filters => {
       return filters
-        .filter(operations => operations.map !== 'c')
+        .filter(operations => operations.map !== MAP_CATEGORY_CHAR)
         .concat(newCategories)
     })
   }
@@ -106,9 +122,19 @@ const FilterSidebar = ({
     const { query, map } = filterContext
     const fullText = getFullText(query, map)
 
+    /* This removes the previously selected stuff from the context when you click on 'clear'.
+    It is important to notice that it keeps categories, departments and clusterIds since they
+    are important to show the correct facets. */
+    if (shouldClear.current) {
+      shouldClear.current = false
+      return filterCategoryDepartmentAndCollection(filterContext)
+    }
+
+    /* The spread on selectedFilters was necessary because buildNewQueryMap
+     changes the object but we do not want that on mobile */
     return {
       ...filterContext,
-      ...buildNewQueryMap(fullText, filterOperations, selectedFilters),
+      ...buildNewQueryMap(fullText, filterOperations, [...selectedFilters]),
     }
   }, [filterOperations, filterContext, selectedFilters])