+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- "Clear" button on mobile filters to uncheck every brand and specification filter.
 
 ## [3.61.0] - 2020-06-19
 
@@ -18,7 +20,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Problem with `useFetchMore` that displayed repeated products.
 
 ## [3.60.5] - 2020-06-15
-
 ### Added
 - Accept `facets` breadcrumb as fallback.